backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.7
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==9.8.0
Markdown==3.3.4
Pillow==8.4.0
# psycopg2==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
django-cors-headers==3.10.0
django-better-admin-arrayfield==1.4.2
ipython
ipdb
